requirements.txt
python-docx>=1.1.0
pandas>=2.0.0
faiss-cpu>=1.7.4
tiktoken>=0.6.0
openai>=1.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
numpy>=1.26.0
pyarrow>=15.0.0
gradio>=4.0.0
